# Conflicts:
#   pom.xml

+++ src/main/java/io/choerodon/devops/app/service/impl/DevopsCdPipelineServiceImpl.java
@@ -493,7 +493,7 @@ public class DevopsCdPipelineServiceImpl implements DevopsCdPipelineService {
                     });
         } catch (Exception e) {
             LOGGER.error("error.create.pipeline.auto.deploy.instance", e);
-            sendFailedSiteMessage(pipelineRecordId, GitUserNameUtil.getUserId().longValue());
+            sendFailedSiteMessage(pipelineRecordId, GitUserNameUtil.getUserId());
             devopsCdStageRecordService.updateStageStatusFailed(stageRecordId);
             devopsCdJobRecordService.updateJobStatusFailed(jobRecordId);
             devopsCdPipelineRecordService.updatePipelineStatusFailed(pipelineRecordId, e.getMessage());
